Add ArchetypeMatch, storeMatch, generate, genTab
Created through storeMatch, ArchetypeMatch stores a match from store
with the transformed parameters and the type of the match. Allowing
ArchetypeMatch#generate to call the Archetype#generate for the
determined Archetype.

genTab creates a tab for the given depth argument to generate to help
printing the generated object with correct tabs.